refactor: consolidate booleans in Team.tsx and organise Sort/Filter options (#4301)
